Removed one unneeded posts controller. Namespace move for other controllers
